Transform VideoControls as an Interface
* Update library: make @EMVideoView@ expected an Interface for @videoControls@ parameters, allowing those to be totally custom
* Update demo according to library modifications
