+++ library/src/main/java/com/devbrackets/android/exomedia/listener/VideoControlsVisibilityListener.java
@@ -18,7 +18,7 @@ package com.devbrackets.android.exomedia.listener;
 
 /**
  * Interface definition for a callback to be invoked when the visibility
- * of the {@link com.devbrackets.android.exomedia.ui.widget.VideoControls} changes
+ * of the {@link com.devbrackets.android.exomedia.ui.widget.BaseVideoControls} changes
  */
 public interface VideoControlsVisibilityListener {
     /**
