+++ library/src/main/java/com/devbrackets/android/exomedia/ui/animation/BottomViewHideShowAnimation.java
@@ -29,7 +29,7 @@ import android.view.animation.AnimationSet;
 import android.view.animation.TranslateAnimation;
 
 /**
- * An animation used to slide {@link com.devbrackets.android.exomedia.ui.widget.VideoControls}
+ * An animation used to slide {@link com.devbrackets.android.exomedia.ui.widget.BaseVideoControls}
  * in and out from the bottom of the screen when changing visibilities.
  */
 public class BottomViewHideShowAnimation extends AnimationSet {
